backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pandas==2.2.0
openpyxl==3.1.2
fpdf2==2.7.7
google-api-python-client==2.116.0
bcrypt==4.1.2
PyJWT==2.8.0
